feat: use memo for task cards
